+++ flask_app/eb-flask/.ebignore
@@ -1,0 +1,1 @@
+virt
