+++ flask_app/eb-flask/requirements.txt
@@ -1,0 +1,13 @@
+boto3==1.17.62
+botocore==1.20.62
+click==7.1.2
+Flask==1.1.2
+itsdangerous==1.1.0
+Jinja2==2.11.3
+jmespath==0.10.0
+MarkupSafe==1.1.1
+python-dateutil==2.8.1
+s3transfer==0.4.2
+six==1.15.0
+urllib3==1.26.4
+Werkzeug==1.0.1
