+++ flask_app/eb-flask/templates/index.html
@@ -21,6 +21,12 @@
                 <p> {{ fname }} </p>
               </li>
           </ul>
+          <h3>Output from Lambda Function in AWS cloud:</h3>
+          <ul>
+              <li>
+                <p> {{ awslambda }} </p>
+              </li>
+          </ul>
         </div>
     </div>
   </body>
