+++ flask_app/legacy_app/templates/index.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Top-n Music Genre Predicter</title>
+  </head>
+  <body>
+    <div class="content">
+        <h3>Music Predicter Demo</h3>
+        <p>This is only a demo of AWS implementation</p>
+        <div>
+          <h3>Upload mp3 file:</h3>
+          <form method="POST" action="/upload" enctype=multipart/form-data>
+            <input type=file name=file>
+            <input type=submit value=Upload>
+          </form>
+        </div>
+        <div>
+          <h3>This file has been uploaded:</h3>
+          <ul>
+              <li>
+                <p> {{ fname }} </p>
+              </li>
+          </ul>
+          <h3>Output from Lambda Function in AWS cloud:</h3>
+          <ul>
+              <li>
+                <p> {{ awslambda }} </p>
+              </li>
+          </ul>
+        </div>
+    </div>
+  </body>
+</html>
